enabled songs link of dashboard so now user can see all songs of platform
